create util file for generate HTML
